Move fallabck generator related code out of fingerprint_suite
