+++ src/crawlee/http_clients/_fallback_consts.py
@@ -1,7 +1,7 @@
 # ruff: noqa: E501
 
 COMMON_ACCEPT = 'text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7'
-
+COMMON_ACCEPT_LANGUAGE = 'en-US,en;q=0.9'
 # Playwright default headers (user-agents and sec-ch) for headless browsers.
 PW_CHROMIUM_HEADLESS_DEFAULT_USER_AGENT = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/124.0.0.0 Safari/537.36'
 PW_CHROMIUM_HEADLESS_DEFAULT_SEC_CH_UA = '"Not=A?Brand";v="8", "Chromium";v="124", "Google Chrome";v="124"'
