+++ src/crawlee/http_clients/_httpx.py
@@ -10,8 +10,8 @@ from crawlee._types import HttpHeaders
 from crawlee._utils.blocked import ROTATE_PROXY_ERRORS
 from crawlee._utils.docs import docs_group
 from crawlee.errors import ProxyError
-from crawlee.fingerprint_suite._fallback_header_generator import HeaderGenerator
 from crawlee.http_clients import HttpClient, HttpCrawlingResult, HttpResponse
+from crawlee.http_clients._fallback_header_generator import HeaderGenerator
 from crawlee.sessions import Session
 
 if TYPE_CHECKING:
